requirements.txt
streamlit
numpy-stl
matplotlib
librosa
voxelfuse